requirements/base.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
idna==3.3
lxml==4.8.0
python-dotenv==0.20.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
SQLAlchemy==1.4.39
urllib3==1.26.9
xsdata==22.4
